requirements.txt
kaggle
matplotlib
ultralytics
pandas
opencv-python
flask
numpy
onnxruntime
pillow
requests
